+++ tests/testthat/test-zzz.R
@@ -1,10 +1,10 @@
-test_that("tests for zzz", {
-  # We need to use callr to avoid conflicts with other tests
-  res <- callr::r(
-    function() {
-      library(linelist)
-      get_lost_tags_action()
-    }
+test_that("tag action is initialized correctly", {
+
+  res <- withr::with_package(
+    "linelist",
+    get_lost_tags_action()
   )
+
   expect_identical(res, "warning")
+
 })